delete BaseEntity
