+++ src/main/kotlin/com/spring_boot/base/model/entity/AbstractEntity.kt
@@ -1,22 +1,34 @@
 package com.spring_boot.base.model.entity
 
+import com.spring_boot.base.model.value_object.AbstractValueObjectId
 import io.swagger.annotations.ApiModelProperty
 import java.time.LocalDateTime
-import javax.persistence.Column
-import javax.persistence.MappedSuperclass
-import javax.persistence.PrePersist
-import javax.persistence.PreUpdate
+import javax.persistence.*
 
 /**
  * abstract class implementing necessary information for Value Object
  */
 @MappedSuperclass
-abstract class AbstractEntity : BaseEntity() {
+abstract class AbstractEntity<T : AbstractValueObjectId> {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @ApiModelProperty(value = "entity ID", required = false)
+    protected var id: T? = null
 
     /**
      * @return [id] value of the entity
      */
-    fun id() = this.id
+    fun id(): T? = this.id
+
+    /**
+     * @return true if the entity is saved, false if not
+     */
+    fun isSaved(): Boolean {
+        return id != null
+    }
+
+    // --------------------------------------
 
     @ApiModelProperty(value = "when it was created", required = false)
     @Column(name = "created_date", nullable = false)
@@ -41,5 +53,36 @@ abstract class AbstractEntity : BaseEntity() {
     @PreUpdate
     protected fun preUpdate() {
         this.updatedDate = LocalDateTime.now()
+    }
+
+    // --------------------------------------
+
+    /**
+     * @return true if [other] is referring the same Entity, false if not
+     */
+    override fun equals(other: Any?): Boolean {
+        if (other === this) return true
+        if (other == null) return false
+        if (other.javaClass != this.javaClass) return false
+        if (other is AbstractEntity<*>) return sameIdentityAs(other)
+
+        return false
+    }
+
+    /**
+     * @return true if [other] has the same id as this entity
+     */
+    protected fun sameIdentityAs(other: AbstractEntity<*>): Boolean {
+        if (id == null || other.id == null) {
+            return false
+        }
+        return this.id == other.id
+    }
+
+    /**
+     * @return hashed code of [id]
+     */
+    override fun hashCode(): Int {
+        return if (id != null) id.hashCode() else 0
     }
 }
