+++ src/main/kotlin/com/spring_boot/base/model/value_object/AbstractValueObjectId.kt
@@ -1,0 +1,5 @@
+package com.spring_boot.base.model.value_object
+
+abstract class  AbstractValueObjectId {
+    abstract fun toEntity()
+}
